Add a missing xml:id attribute to <msItem> nodes
This attribute is how we identify parts of a manuscript in the front-end,
and they're used to create the identifiers that go in works URLs
(e.g. /works/qxvg3rvp)

The front-end is warning that these attributes are missing, so I've added
them; this should allow these manuscripts to appear online.